+++ colormaps/managers.py
@@ -11,7 +11,9 @@ import os
 
 from colormaps import core
 
-PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), 'data'))
+BASE_PATH = os.path.dirname(__file__)
+DATA_PATH = os.path.abspath(os.path.join(BASE_PATH, 'data'))
+LABELS_PATH = os.path.abspath(os.path.join(BASE_PATH, 'labels'))
 
 
 class Manager(object):
@@ -25,7 +27,21 @@ class Manager(object):
     def get_id(self, path):
         return os.path.splitext(os.path.basename(path))[0]
 
-    def __init__(self, path=PATH):
+    def load_label(self, path):
+        with open(path) as json_file:
+            return dict(json.load(json_file))
+
+    def get_labels(self, name):
+        """ Return locale dict of label dicts. """
+        base = os.path.join(LABELS_PATH, name)
+        if not os.path.exists(base):
+            return {}
+        names = os.listdir(base)
+        paths = [os.path.join(base, n) for n in names]
+        locales = [os.path.splitext(n)[0] for n in names]
+        return {l: self.load_label(p) for l, p in zip(locales, paths)}
+
+    def __init__(self, path=DATA_PATH):
         self.path = path
 
         # read data
@@ -46,7 +62,9 @@ class Manager(object):
         if name in core.registered:
             return core.registered[name]
         if name in self.registered:
-            colormap = core.create(json.load(open(self.registered[name])))
+            config = json.load(open(self.registered[name]))
+            config['labels'] = self.get_labels(name)
+            colormap = core.create(config)
             colormap.register(name)
             return colormap
         raise NameError("'{}' is not in registered colormaps".format(name))